消息机制重构: 消除 InputListView 与 InputList 之间的绑定

+++ app/src/main/java/org/crazydan/studio/app/ime/kuaizi/pane/view/KeyboardViewBase.java
@@ -103,7 +103,7 @@ public abstract class KeyboardViewBase extends RecyclerView {
         this.adapter.updateKeys(keys, themeResId, orientation);
     }
 
-    public double getBottomSpacing() {
+    public float getBottomSpacing() {
         return this.layoutManager.getGridPaddingBottom();
     }
