+++ app/src/main/java/org/crazydan/studio/app/ime/kuaizi/pane/view/key/KeyViewLayoutManager.java
@@ -246,7 +246,7 @@ public class KeyViewLayoutManager extends RecyclerViewLayoutManager {
         return rectHexagonList;
     }
 
-    public double getGridPaddingBottom() {
+    public float getGridPaddingBottom() {
         return this.gridPaddingBottom;
     }
